hung bui commit admin web app

+++ Source/JPA_Spring/fparking/src/main/java/com/tagroup/fparking/service/domain/Driver.java
@@ -1,15 +1,23 @@
 package com.tagroup.fparking.service.domain;
 
 import java.io.Serializable;
-import javax.persistence.*;
-
-import org.hibernate.annotations.Proxy;
-
 
 //import com.ctwgroup.erestaurant.service.domain.Category;
 //import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import java.util.List;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.NamedQuery;
+import javax.persistence.Table;
+
+import org.hibernate.annotations.Proxy;
 
 
 /**
@@ -25,7 +33,7 @@ public class Driver implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 	
 	private String name;
 	
@@ -41,11 +49,11 @@ public class Driver implements Serializable {
 	public Driver() {
 	}
 
-	public int getId() {
+	public Long getId() {
 		return this.id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
